add time_util.rb
